refactor: get post data
use mixin(post)
