+++ enhanceApp.js
@@ -1,0 +1,7 @@
+import postMixin from '@theme/mixins/posts'
+
+export default ({
+  Vue
+}) => {
+  Vue.mixin(postMixin)
+}
